build with openmpi now works.
The build script used to detect the local installation and copy the setup
file to the installation directory, regardless of whether the local installation
were being used. This problem is now resolved and the setup file is copied to the
installation directory only when the local MPI library or compilers are used to
build the library.