stack to heap bug fix
This commit resolves the bug that was introduced when attempting to silence
the GFortran warning message:

...moved from stack to static storage. This makes the procedure unsafe when called recursively,
or concurrently from multiple threads. Consider using '-frecursive'...

The -frecursive flag has the potential to generate segfaults and in this particular case,
resulted in deadlocks in the Coarray builds of the ParaMonte library.

Based on the community's opinion, as long as the recursive functions are not used or
are carefully designed, this warning message can be safely ignored.

Also, keep in mind that heap arrays are 100X or more slower than stack allocations.