test build bug fix + naming convention
A bug that prevented the test build in the shared library build mode is now resolved.
The bug originated from renaming the shared libraries from DYNAMIC to SHARED, while the test CMakeFile
still required and conditioned on the keyword DYNAMIC.

The naming conventions for the file names and modules are now more consistent with other files,
in particular, in the light of the new ParaNest code additions.