ProposalErr is now removed from the ParaDRAM ProposalAbstract class.
ProposalErr is now part of the Proposal_type class in the Proposal modules
of the ParaDRAM sampler. Consequently, the methods getNew() and doAdaptation()
are now dynamic methods rather than static.

My performance benchmarks indicate that no change has occurred in the performance of the code.
This is one more step toward making ParaMonte thread-safe.