major updates to MATLAB visualization tools.
This is a major syntax breaking update to the MATLAB vis tools of the ParaMonte library.
Specifically, we had to make the vis function and class names unique so that the
Doxygen documentater could separate the docs for each class.

Therefore, now for example, instead of,

```
pm.vis.plot.Line()
pm.vis.tile.Line()
pm.vis.subplot.Line()
pm.vis.cascade.Line()
```

we have

```
pm.vis.PlotLine()
pm.vis.TileLine()
pm.vis.SubplotLine()
pm.vis.CascadeLine()
```

The documentation of all vis functions are nearly complete,
except for pm.vis.Tile* as well as pm.visCorner* which is yet to be completed.

Additionally, the ``dims`` component of the Ellipse classes are split into
two separate components ``dimx`` and ``dimy`` to facilitate easy assignment of
multiple columns on the X or Y axes, without having to create a matrix for ``dims``.

Additionally, the ``colx`` and ``rows`` components of the CascadeHeatmap class
can take input cell arrays, each element of which will be assigned
to a separate figure in the cascade of figures.